// src/app/movies-list/model/filter.ts
export enum Filter {
    POPULAR_DESC = 'Popularitet fallande',
    POPULAR_ASC = 'Popularitet stigande',
    VOTE_DESC = 'Betyg fallande',
    VOTE_ASC = 'Betyg stigande',
    RELEASE_DATE_DESC = 'Utgivningsdatum fallande',
    RELEASE_DATE_ASC = 'Utgivningsdatum stigande',
    TITLE_DESC = 'Titel (A-Ö)',
    TITLE_ASC = 'Titel (Ö-A)'
}